Add files via upload
Code for rank-reduced envelope regression. This codes only estimates the beta so far. It prints an the Frobenius norm between RR-Env, RR, Env and OLS. I eye balled the results and it looks that the plots will be reproducible with this code. At least in behavior, may be off in actual values (which is not important). It uses the files obtained from data_simulation.cpp
It is fairly commented if you have questions let me know.